Fix validation crash and implement performance optimizations
- Fix RMSNorm bfloat16 memory bug by implementing custom RMSNorm
- Add true batching support to DocumentStreamWrapper
- Update training loop to handle batched inputs with dynamic state resets
- Fix validation code to work with new model signature
- Add torch.compile support for performance
- Optimize gossip serialization to use bfloat16
- Reduce expansion_factor from 4.0 to 2.0 for memory savings
- Set chunk_size to 512 and batch_size to 8
- Add --bf16 flag to training script

+++ gossip/evolutionary_node.py
@@ -887,7 +887,7 @@ class EvolutionaryTrainingNode:
                 'validation_fitness': self.get_current_fitness(),
                 'ema_loss': self.get_current_fitness(),  # Keep for compatibility
                 'model_state_dict': {
-                    name: param.detach().cpu() 
+                    name: param.detach().cpu().to(torch.bfloat16) 
                     for name, param in self.model.named_parameters()
                 }
             }
